Made parametric pulley gear.
Modified version of [the customizable pulley generator on
Thingiverse](http://www.thingiverse.com/thing:16627), this one includes
a slight modification to provide a nub on the end of the pulley, for
connecting to an idler bearing.

I'd previously had the old pulley generator checked in here, I've now
removed all the unrelated files, and moved the customizer into
`pulley.scad`.